+++ src/index.ts
@@ -1,15 +1,15 @@
 import Engine from './Engine.ts';
-
 import Sprite, { type SpriteOptions } from './Sprite.ts';
+import { setScale, setAspectRatio, canvas, ctx } from './canvas.ts';
+
 import Rectangle, { type RectangleOptions } from './sprites/Rectangle.ts';
+import Square, { type SquareOptions } from './sprites/Square.ts';
 import Oval, { type OvalOptions } from './sprites/Oval.ts';
 import Circle, { type CircleOptions } from './sprites/Circle.ts';
 import RegularPolygon, { type RegularPolygonOptions } from './sprites/RegularPolygon.ts';
 import Pen, { type PenOptions } from './sprites/Pen.ts';
 import Text, { type TextOptions, type CanvasTextAlign, type CanvasTextBaseline } from './sprites/Text.ts';
 import ImageSprite, { type ImageSpriteOptions } from './sprites/ImageSprite.ts';
-
-import { setScale, setAspectRatio, canvas, ctx } from './canvas.ts';
 
 const TScratch = {
     // Main
@@ -18,6 +18,7 @@ const TScratch = {
 
     // Sprites
     Rectangle,
+    Square,
     Oval,
     Circle,
     RegularPolygon,
@@ -40,6 +41,7 @@ export {
 
     // Sprites
     Rectangle,
+    Square,
     Oval,
     Circle,
     RegularPolygon,
@@ -50,6 +52,7 @@ export {
     // Options
     type SpriteOptions,
     type RectangleOptions,
+    type SquareOptions,
     type OvalOptions,
     type CircleOptions,
     type RegularPolygonOptions,